Refactor: Nav바 UI 리팩토링
Nav 컴포넌트 파일  Main으로 이동, index.js 수정

Issue: #2
