+++ src/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import Router from "./Router";
-import Navigation from "./components/Nav/NavigationBar";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(<Navigation />);
+root.render(<Router />);
