+++ src/pages/Main/Main.js
@@ -1,7 +1,12 @@
 import styles from "./Main.module.scss";
+import Navigation from "./Nav/NavigationBar";
 
 function Main() {
-  return <div className={styles.container}>Main</div>;
+  return (
+    <div>
+      <Navigation />
+    </div>
+  );
 }
 
 export default Main;
