+++ src/pages/Main/Nav/DropDownMenu/WorkMenu.js
@@ -1,9 +1,9 @@
 import styles from "./DropDownMenu.module.scss";
-import carpentry from "../../../assets/images/carpentry.png";
-import leather from "../../../assets/images/leather.png";
-import jewelry from "../../../assets/images/jewelry.png";
-import pottery from "../../../assets/images/pottery.png";
-import tools from "../../../assets/images/tools.png";
+import carpentry from "../../../../assets/images/carpentry.png";
+import leather from "../../../../assets/images/leather.png";
+import jewelry from "../../../../assets/images/jewelry.png";
+import pottery from "../../../../assets/images/pottery.png";
+import tools from "../../../../assets/images/tools.png";
 
 function WorkMenu({ workMenuOpen }) {
   return (
